fix migrations

+++ db/migrate/20091202204807_reference_number_types_add_default_column.rb
@@ -1,9 +1,13 @@
 class ReferenceNumberTypesAddDefaultColumn < ActiveRecord::Migration
-  	def self.up
+  def self.up
+    #Lht follow pre-migration
+    unless column_exists? :reference_number_types, :default
+      remove_column :reference_number_types, :default
   		add_column :reference_number_types, :default, :boolean, :default => false
-  	end
+    end
+  end
 
-  	def self.down
-  		remove_column :reference_number_types, :default
-  	end
+  def self.down
+    remove_column :reference_number_types, :default
+  end
 end
